+++ source/convertToJson.js
@@ -51,6 +51,9 @@ export default function(data, schema, options) {
   // Correct error rows.
   if (rowMap) {
     for (const error of errors) {
+      // Convert the `row` index in `data` to the
+      // actual `row` index in the spreadsheet.
+      // The `1` compensates for the header row.
       error.row = rowMap[error.row] + 1
     }
   }
@@ -192,7 +195,6 @@ function parseValueOfType(value, type, options) {
       return { value }
 
     case Number:
-    case 'Integer':
     case Integer:
       // The global isFinite() function determines
       // whether the passed value is a finite number.
@@ -211,14 +213,12 @@ function parseValueOfType(value, type, options) {
       }
       return { value }
 
-    case 'URL':
     case URL:
       if (!isURL(value)) {
         return { error: 'invalid' }
       }
       return { value }
 
-    case 'Email':
     case Email:
       if (!isEmail(value)) {
         return { error: 'invalid' }