+++ source/readXlsx.test.js
@@ -1,6 +1,26 @@
 import { dropEmptyRows, dropEmptyColumns } from './readXlsx'
 
 describe('readXlsx', () => {
+	it('should drop empty rows (only at the end)', () => {
+		dropEmptyRows([
+			[null, null, null],
+			['A', 'B', 'C'],
+			[null, 'D', null],
+			[null, null, null],
+			['E', 'F', 'G'],
+			[null, null, null]
+		], {
+			onlyTrimAtTheEnd: true
+		})
+		.should.deep.equal([
+			[null, null, null],
+			['A', 'B', 'C'],
+			[null, 'D', null],
+			[null, null, null],
+			['E', 'F', 'G']
+		])
+	})
+
 	it('should drop empty rows', () => {
 		dropEmptyRows([
 			[null, null, null],
@@ -14,6 +34,23 @@ describe('readXlsx', () => {
 			['A', 'B', 'C'],
 			[null, 'D', null],
 			['E', 'F', 'G']
+		])
+	})
+
+	it('should drop empty columns (only at the end)', () => {
+		dropEmptyColumns([
+			[null, 'A', 'B', 'C', null, null],
+			[null, 'D', null, null, null, null],
+			[null, null, null, null, null, null],
+			[null, null, 'E', 'F', 'G', null]
+		], {
+			onlyTrimAtTheEnd: true
+		})
+		.should.deep.equal([
+			[null, 'A', 'B', 'C', null],
+			[null, 'D', null, null, null],
+			[null, null, null, null, null],
+			[null, null, 'E', 'F', 'G']
 		])
 	})
 
@@ -33,7 +70,7 @@ describe('readXlsx', () => {
 	})
 
 	it('should generate row map when dropping empty rows', () => {
-		const rowMap = []
+		const rowMap = [0, 1, 2, 3, 4]
 
 		dropEmptyRows([
 			[null, null, null],
@@ -42,7 +79,7 @@ describe('readXlsx', () => {
 			[null, null, null],
 			['E', 'F', 'G']
 		],
-		rowMap)
+		{ rowMap })
 		.should.deep.equal([
 			['A', 'B', 'C'],
 			[null, 'D', null],